fix bug

+++ berkeley-function-call-leaderboard/model_handler/toolace_handler.py
@@ -17,7 +17,7 @@ NO other text MUST be included.
 
 class ToolACEHandler(OSSHandler):
     def __init__(self, model_name, temperature=0.001, top_p=1, max_tokens=1000) -> None:
-        super().__init__(model_name, temperature, top_p, max_tokens, dtype=dtype)
+        super().__init__(model_name, temperature, top_p, max_tokens)
 
     def _format_prompt(prompts, function, test_category):
